pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-docs-normalizer"
version = "0.1.0"
description = "Herramienta CLI para normalizar documentación de APIs mal estructurada a OpenAPI 3.0"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Luis Ruiz de Arbulo"}
]
keywords = ["openapi", "api", "documentation", "normalizer", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
api-normalizer = "api_normalizer.cli:main"

[tool.setuptools]
packages = ["api_normalizer", "api_normalizer.exporters"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]
